clean up gluon contrib
