+++ docs/python_docs/python/api/gluon/data/index.rst
@@ -39,6 +39,7 @@ Sampling
    SequentialSampler
    RandomSampler
    BatchSampler
+   IntervalSampler
 
 DataLoader
 ----------
