+++ python/mxnet/contrib/__init__.py
@@ -24,7 +24,6 @@ from . import ndarray
 from . import symbol as sym
 from . import ndarray as nd
 
-from . import autograd
 from . import tensorboard
 
 from . import text
