+++ python/mxnet/gluon/contrib/__init__.py
@@ -18,12 +18,6 @@
 # coding: utf-8
 """Contrib neural network module."""
 
-from . import nn
-
-from . import rnn
-
-from . import cnn
-
 from . import data
 
 from . import estimator
