+++ python/mxnet/gluon/contrib/data/__init__.py
@@ -19,7 +19,4 @@
 # pylint: disable=wildcard-import
 """Contrib datasets."""
 
-from . import text
 from . import vision
-
-from .sampler import *
