+++ python/mxnet/gluon/model_zoo/vision/squeezenet.py
@@ -25,7 +25,6 @@ import os
 from ....context import cpu
 from ...block import HybridBlock
 from ... import nn
-from ...contrib.nn import HybridConcurrent
 from .... import base
 
 # Helpers
@@ -33,7 +32,7 @@ def _make_fire(squeeze_channels, expand1x1_channels, expand3x3_channels):
     out = nn.HybridSequential()
     out.add(_make_fire_conv(squeeze_channels, 1))
 
-    paths = HybridConcurrent(axis=1)
+    paths = nn.HybridConcatenate(axis=1)
     paths.add(_make_fire_conv(expand1x1_channels, 1))
     paths.add(_make_fire_conv(expand3x3_channels, 3, 1))
     out.add(paths)
