+++ python/mxnet/numpy/random.py
@@ -18,13 +18,15 @@
 """Namespace for ops used in imperative programming."""
 
 from ..ndarray import numpy as _mx_nd_np
+from ..random import seed
 
 
 __all__ = ["randint", "uniform", "normal", "choice", "rand", "multinomial", "multivariate_normal",
            "logistic", "gumbel", "f",
            "laplace",
            "shuffle", "randn", "gamma", "beta", "chisquare", "exponential", "lognormal",
-           "weibull", "pareto", "power", "rayleigh"]
+           "weibull", "pareto", "power", "rayleigh",
+           "seed"]
 
 
 def randint(low, high=None, size=None, dtype=None, ctx=None, out=None):
